chore: 상수파일 분리
