+++ src/constants/Follower/currentConst.ts
@@ -1,0 +1,40 @@
+export const GROUPS = ['1번 스터디', '2번 스터디', '3번 스터디', '4번 스터디'];
+
+export const SORTING = ['최신순', '|', '가나다순'];
+
+export const DUMMY = {
+  profile: {
+    imgSrc: 'https://avatars.githubusercontent.com/u/80264647?v=4',
+    nickname: '닉네임은총열글자입니',
+  },
+  boards: [
+    {
+      date: '26',
+      count: 1,
+    },
+    {
+      date: '27',
+      count: 0,
+    },
+    {
+      date: '28',
+      count: 0,
+    },
+    {
+      date: '29',
+      count: 0,
+    },
+    {
+      date: '30',
+      count: 0,
+    },
+    {
+      date: '31',
+      count: 0,
+    },
+    {
+      date: '1',
+      count: 0,
+    },
+  ],
+};
