--- src/constants/FollowerCurrent/currentConst.ts
@@ -1,8 +1,0 @@
-export const GROUPS = [
-  '1번 스터디',
-  '2번 스터디',
-  '3번 스터디',
-  '4번 스터디',
-];
-
-export const SORTING = ['최신순', '|', '가나다순'];
