+++ src/page/FollowerPage.tsx
@@ -2,106 +2,7 @@ import styled from 'styled-components';
 import FollowerRecommendCard from '../components/Follower/Personal/FollowerRecommendCard';
 import ParticipatingGroup from '../components/Follower/Personal/ParticipatingGroup';
 import PageLayout from '../components/PageLayout/PageLayout';
-
-const DUMMY = {
-  profileImg: '',
-  nickname: '코딩하는 갱얼쥐',
-  isFollowed: false,
-  introduce:
-    '나를 소개하는 문구를 적어주세요.아아아 최대 두줄까지만 적을 수 있게 하는거어떠나를 소개하는 문구를 적어주세요.아아아 최대 두줄까지만 적을 수 있게 하는거어떠',
-  github: 'https://github.com/Arooming',
-  group: [
-    {
-      id: 0,
-      imgSrc:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRgQdMSF0UfUeev25_EDlcSS0jCfciYTLT-qw&s',
-      title: '이제는 더이상 물러날 곳이 없는 스터디',
-      tags: ['#swift', '#java', '#javascript'],
-      introduce:
-        '안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다',
-    },
-    {
-      id: 1,
-      imgSrc:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRolslbPtm3OymbdIwYgusZ9wqH4-fvxhln2A&s',
-      title: '이제는 더이상 물러날 곳이 없는 스터디',
-      tags: ['#swift', '#java', '#javascript'],
-      introduce:
-        '안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다',
-    },
-    {
-      id: 2,
-      imgSrc:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcTwwTi48CyZZGovLP7-u6CPqTn8ygolt2N5aA&s',
-      title: '이제는 더이상 물러날 곳이 없는 스터디',
-      tags: ['#swift', '#java', '#javascript'],
-      introduce:
-        '안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다',
-    },
-    {
-      id: 3,
-      imgSrc:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRgQdMSF0UfUeev25_EDlcSS0jCfciYTLT-qw&s',
-      title: '이제는 더이상 물러날 곳이 없는 스터디',
-      tags: ['#swift', '#java', '#javascript'],
-      introduce:
-        '안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다',
-    },
-    {
-      id: 4,
-      imgSrc:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcRolslbPtm3OymbdIwYgusZ9wqH4-fvxhln2A&s',
-      title: '이제는 더이상 물러날 곳이 없는 스터디',
-      tags: ['#swift', '#java', '#javascript'],
-      introduce:
-        '안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다설명칸은 두줄까지 가능합니다 안녕하세요 스터디입니다 설명칸은 두줄까지 가능합니다',
-    },
-  ],
-  recommend: [
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐1',
-      language: 'javascript',
-      isFollowed: false,
-    },
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐2',
-      language: 'javascript',
-      isFollowed: false,
-    },
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐3',
-      language: 'javascript',
-      isFollowed: false,
-    },
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐4',
-      language: 'javascript',
-      isFollowed: false,
-    },
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐5',
-      language: 'javascript',
-      isFollowed: false,
-    },
-    {
-      profileImg:
-        'https://encrypted-tbn0.gstatic.com/images?q=tbn:ANd9GcQflIJYye-3WuwXfwuKLvVcajyyi8Rbu9Bx0g&s',
-      nickname: '코딩하는 갱얼쥐6',
-      language: 'javascript',
-      isFollowed: false,
-    },
-  ],
-};
+import { FOLLOWER_DUMMY } from '../constants/Follower/followerConst';
 
 const FollowerPage = () => {
   const {
@@ -112,7 +13,7 @@ const FollowerPage = () => {
     // github,
     group,
     recommend,
-  } = DUMMY;
+  } = FOLLOWER_DUMMY;
   return (
     <PageLayout category="홈">
       <FollowerPageContainer>
